Revert r107440 because this causes a crash on touchui on page load.
The CL being reverted:
This patch lets the browser window punch a hole in the DesktopBackgroundView,
thus eliminating overdraw.
To do this, the hole punching logic was changed to look at the entire tree.
Drawing the screen is a preorder traversal through all the layers. Thus, to
compute the hole, we have to look at all of the layers after a particular layer
in the preorder traversal.
This is not an efficient algorithm, but it is simplest I could come up with.
Plus we will be moving to the WebKit compositor soon.

Hopefully this will improve the frame rate for the Aura demos

Note: I will put the changes to Transform into another patch.

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=107360

Review URL: http://codereview.chromium.org/8368013

TBR=pkotwicz@chromium.org

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@108347 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 02876f5fdafb3b6e500ba322715c3a848170763a