Fix stale compositor references from ui::Layer
This is needed to fix some issues in tests exposed by the WebKit compositor. The
problem is that sometimes during the teardown paths, some layers have stale a
pointer to the compositor that has already been destroyed. The WebKit layer
(rightfully) calls ScheduleDraw when the layer hierarchy is changed, exposing
the issue.

We now always walk back to the root layer to find the compositor, and reset the
pointer when the compositor's root changes (or it gets destroyed).

BUG=99524
TEST=views_unittest, aura_unittest, compositor_unittests with (and without) use_webkit_compositor=1

Review URL: http://codereview.chromium.org/8510076

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@110223 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ab3f2755bd71f7b7a9fbed5fef4c3bec628fda6