Render Core Animation plugins through WebKit's compositor rather than directly to the screen in the browser process.
The new composited code path is now the default, though the old code path
has been left in place under a command line flag while we gain confidence.
Issue 105344 has been filed about removing the old code path.

The new code path does not currently support 10.5. The consequence is that
plugins using the InvalidatingCoreAnimation rendering model will not work
on this version of Mac OS. Pepper 3D is not affected; it now uses a
different rendering path.

Changed the type of IOSurfaces' IDs from uint64 to uint32 in a few places
throughout the code to match the IOSurfaceID typedef in the system header.
This was necessary in order to simplify integration with Chrome's OpenGL
code.

There is a known problem in the new code path with garbage occasionally
being drawn to the plugin's area during live resizing of Core Animation
plugins. Issue 105346 has been filed to track this. It is unclear whether
the additional complexity of the fix that is likely needed is worth it.

Tested manually with the following content, with and without the
--disable-composited-core-animation-plugins flag:
 - YouTube (does not trigger this code path)
 - Google+ Hangouts
 - http://unity3d.com/gallery/demos/live-demos (Unity 3D)
 - http://www.erain.com/labs/molehill/ (Stage 3D in Flash 11)
 - http://www.nissan-stagejuk3d.com/
    (Stage 3D in Flash 11, live resizing; web site is flaky, sometimes
    fails to start)

BUG=38967
TEST=manual testing with above test cases

Review URL: http://codereview.chromium.org/8678037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@112126 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc51d1ccc18135493f01464e38ed462825ae8e35