Revert 112770 - Prevent a crash in the first run search engine selector.
It is possible to try and close the window now before the TemplateURLService is loaded. In this case we will not have a fallback search engine to select by default. So we will just reject attempts to close the window in this circumstance until it is loaded and a fallback choice has been set. It should be an edge case.

http://crbug.com/106078
TEST=unit test
Review URL: http://codereview.chromium.org/8771018

TBR=ben@chromium.org
Review URL: http://codereview.chromium.org/8790001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@112804 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db84ed0319cadfcff26cf9892f7fef2a84156bfb