aura: Ask the WindowDelegate before a Window is Focus()ed.
. In aura, before a window is focused, ask the delegate if it can be focused.
. Make a popup RWHVA non-focusable.

This is to prevent the focus from moving from the main web-page to the combobox (which causes the combobox to get destroyed). This fixes an issue where comboboxes in the login screen could not be used properly.

BUG=105786
TEST=manually

Review URL: http://codereview.chromium.org/8770011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@112909 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8bb9670c4e3e91c5f081a272a6d51ed8ee76d8f