Reland 112770, but with test disabled.
Prevent a crash in the first run search engine selector.

It is possible to try and close the window now before the TemplateURLService is loaded. In this case we will not have a fallback search engine to select by default. So we will just reject attempts to close the window in this circumstance until it is loaded and a fallback choice has been set. It should be an edge case.

http://crbug.com/106078
TEST=unit test

Review URL: http://codereview.chromium.org/8771018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@113149 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a25b2fc9f77db19eb6bd2997f3e3ff1c02841538