Don't tell browser to swap until swap actually executed.
By deferring to fence, we're ensuring that the GL commands are executed GPU side before sending a message to the browser to act on them. This is lighter weight than a glFinish because the other commands buffers on the channel (e.g. WebGL) can continue executing while we're waiting for the fence.

TEST=by hand building use_aura running http://webglsamples.googlecode.com/hg/aquarium/aquarium.html

Review URL: http://codereview.chromium.org/8497015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@114798 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d76ba379fef00cbeed525a7911b89400b0ed853