Hide the view which is removed from its parent.
Sometimes the view can receive some events after the remove but before really
deleted and it's not safe because the view does not belong to the root window
in the meantime.

In this specific test,  the JS test code shows infobar and immediately creates
a new tab, so UpdateRect of the infobar contents and removing of the infobar
happen at the almost same time.  If the event happens in the order of
RemoveFromWidget() -> UpdateRect -> delete, the UpdateRect will cause a crash.

BUG=105177
TEST=passed locally

Review URL: http://codereview.chromium.org/9253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@118273 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 96083ba7da51ea30bd5fb2b317643af77ad1eac5