Reland "aura/chromeos: Avoid suspending while video is playing."
This reverts r118199, which was a revert of r118171.
aura::VideoPropertyWriter (not ash::VideoDetector -- made a
typo in the revert's description) was outliving the
ash::Shell singleton.  Now we explicitly destroy the
VideoPropertyWriter before the shell.

The only differences from r118171 are the addition of the
final chunk in chrome_browser_main_chromeos.cc and a change
to reparent a window in VideoDetectorTest.WindowNotVisible
needed in response to r118259.

BUG=110114
TEST=avoided missing DCHECK on shutdown in trybot logs this time

Review URL: http://codereview.chromium.org/9254018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@118375 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6aa614a9ac971f5b027eee0749f211883f1d6934