Revert 119457 - --test-gl-lib for GpuPixelBrowserTests
This CL causes the GpuPixelBrowserTests to load libllvmpipe.so (if it is available) and fallback to standard libGL.so otherwise. This is so that we can (ultimately) do a smoke test of the GPU accelerated display path on the Aura dbg bot on the Chromium waterfall.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/9169100

TBR=backer@chromium.org
Review URL: https://chromiumcodereview.appspot.com/9225029

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119468 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 42bba4a1a7c8f0443810dd7c983fd52013d5c755