Avoid excessive calls to DeriveFont() in RenderTextWin.
This is a performance optimization. PlatformFontWin::DeriveFont() is
slow because it calls GetTextMetrics() and GetTextExtentPoint32() from
PlatformFontWin::CreateHFontRef().

BUG=105550
TEST=Build Chrome Windows with use_canvas_skia_skia=1. Go to paypal.com
and click on the SSL box in the omnibox. The SSL bubble that is brought
up should be much less janky.

Review URL: http://codereview.chromium.org/9314044

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120353 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: add1cca97a1d5de2f096fc8101f04a5026a97266