Fixes login screen menus.
Pass in an instance of a window to use for testing whether the inner dispatcher can receive events.
The |associated_window| should be such that it is behind a window which stops event propagation whenever the inner dispatcher should not receive events.

This fixes menus on the login screen because menus are above the lock screen in z order.

Bug=113247
Testing=Manual

Review URL: http://codereview.chromium.org/9374028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121570 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32b1640238c621ac9153711c644446b48afe410f