ash: Allow dragging using touch/gesture events.
This is the first step towards making it possible to drag with touch/gesture events. There will be subsequent CLs to make sure the correct touch-id is used for dragging, and update the drag-image position to be slightly above the touch-point so that it doesn't get obscured by the touch-finger.

BUG=114752
TEST=mouse-event dragging should be covered by existing tests. Tests for touch-events will be added in subsequent CLs.

Review URL: https://chromiumcodereview.appspot.com/9418043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@122982 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3a37da5433dced78a29befc72fa69512ff96371