aura: Switch to light hinting for UI text.
The eternal debate continues...

This makes us use light hinting instead of strong hinting
for Pango-Cairo-rendered text in the UI (including the tab
strip and the wrench menu), which improves the shaping of
glyphs like 'B' and 'T'.  On the other hand, it makes color
fringes more noticeable, although the effect is worse on e.g.
my Dell 2405FPW than on Lumpy.  I believe that the long-term
plan is to instead draw this text using Skia and improve
subpixel rendering there.

The change also has the added-but-unintended benefit of
fixing some off-by-one misplacement of glyphs in the omnibox
that I believe we picked up when we switched from GTK+ to
Aura Views.

(I tried medium hinting as well, but it's identical to strong
hinting.)

For before-and-after, see
http://www/~derat/no_crawl/20120309/tabs.png and
http://www/~derat/no_crawl/20120309/menu.png.

BUG=116296,117268
TEST=manual inspection on lumpy

Review URL: http://codereview.chromium.org/9664007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@125906 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4cabe55b01920321bd7183f60058a2887e95da0e