Reland 125850 - Aura: Throttle resize to ability to draw
At the root_window.cc level we defer mouse events triggering a resize while a draw is pending. We are careful to keep the sequence of different event types the same. But we aggregate synthetic mouse moves and mouse drags wwhere possible.

In RWHVA, we ask the root window to throttle the resize if we are waiting on the GPU process for a new image. The net result is that UI can only be one frame ahead of the tab contents that are being resized.

BUG=58782
TEST=goto google.com with --force-compositing-mode; resize the window by stretching to the right; see very little gap between tab contents and window frame

Review URL: https://chromiumcodereview.appspot.com/9662020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126132 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 308aaa31d91a166e8223a89cb84559bc80d2d1bb