Touch Lock is based on proximity to other touches.
The Gesture Recognizer now uses a separate Gesture Sequence for each Window.
It also stores a mapping from touch id to Window.

When a new touch comes in:

If there is a capture window, it is associated with the capture window.

Otherwise, if the Gesture Recognizer already has the touch id associated with a target,
the new touch is associated with the same target.

If not, the GestureRecognizer looks at all other current touches, to see if any is within a
certain radius of the new touch. If there is a nearby touch, the new touch is associated
with the closest touch.

If a target still can't be found, the touch is associated with the window it occurred above.

BUG=116585
TEST=GestureRecognizerTest.GestureEventTouchLockSelectsCorrectWindow

Review URL: http://codereview.chromium.org/9666004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127997 0039d316-1c4b-4281-b951-d872f2087c98

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 728aad6d829f401d0b167366cc5a32197cae1b96